Ficha6/src/test/java/Ex2Test.java
import org.example.Ex2;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Nested;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.ValueSource;

import static org.junit.jupiter.api.Assertions.assertTrue;

public class Ex2Test {

    private Ex2 numeros;

    @BeforeAll
    public static void setupAll() {
        System.out.println("INICIO");
    }

    @BeforeEach
    public void setup() {
        System.out.println("Instantiating Contact Manager");
        numeros = new Ex2();
    }

    @Nested
    class ParameterizedTests {
        @DisplayName("Verificar primos")
        @ParameterizedTest
        @ValueSource(ints = {3, 23, 311, 487, 653, 947})
        public void verificarNumeroPrimo(int number) {
            assertTrue(numeros.isPrime(number));
        }

        @DisplayName("Verificar pares")
        @ParameterizedTest
        @ValueSource(ints = {32, 64, 2, 20, 30, 26})
        public void verificarNumeroPar(int number) {
            assertTrue(numeros.isEven(number));
        }

        @DisplayName("Verificar múltiplos")
        @ParameterizedTest
        @ValueSource(ints = {23, 46, 115, 184, 207, 230})
        public void verificarNumeroMultiplo(int number) {
            assertTrue(numeros.isMultiple(number,23));
        }
    }

}